IMU class implementation
pulled the IMU class into the main code structure and seems to be working. still need to run tests

+++ PDC/PDC_mainCode/PDC_SPI.h
@@ -1,1 +1,1 @@
-unsigned int readSPI(int deviceSelect, byte registerSelect, int numBytes);
+unsigned int readSPI(int deviceSelect, int registerSelect, int numBytes);
